moved LoginSlice to Redux
