+++ src/Redux/store.js
@@ -1,8 +1,7 @@
-import { configureStore } from '@reduxjs/toolkit'
-import loginReducer from '../Components/LoginSlice'
-
+import { configureStore } from "@reduxjs/toolkit";
+import loginReducer from "./LoginSlice";
 export default configureStore({
   reducer: {
     loggedIn: loginReducer,
   },
-})
+});
